Improve logic and move logic from helpers to locator popup
